Fix relative path resolution (#29)
This fixes a bug with resolving links when multiple files have the same name. This would select the first page found in files.src_paths, which might be the wrong one and not contain the right anchor.

To fix, we construct the full expected path based on the file that contained the link.