Provide compatibility with mkdocs-material blog plugin
Prior to this commit, certain assumptions were made about the files
seen in `on_files` that are not true when the mkdocs-material blog
plugin is used.

The url seen at the time `on_files` is called is not guaranteed to
be the final url that will appear in html. With the blog plugin,
at the time `on_files` is called, the value will be something like:

`blog/posts/foo.md`

but at the time we are trying to get a mapping, the url will be
something like:

`blog/2024/01/foo.html`

Due to this change, htmlproofer would fail to validate the url despite
it being valid.

To address this, instead of looking up pages from a Dict where the
key is set at the time of `on_files`, we know store a list of Files
and check for the `search_path` for a url against the `url` attribute
of each File to try to find the mapping.