Update contributing development section (#74)
The instructions didn't get you a working environment for testing.
They also were missing a number checks that happen in CI. This
now documents everything you need to run locally to get CI to pass
except for the "running on different platforms". All the Ubuntu checks
seems mostly as "safety" but the Windows checks can discover bugs that
don't occur on Linux. I haven't documented that at this time as it
was outside of the scope of what I wanted to accomplish with this change.